Attempt to cure Millstone Curse
The Millstone was possessed by a Curse which somehow corrupted the Recipe Registry to reset its Item Count to 0. A Java Ritual of Redundancy and Boilerplate with enough certified Java Black Magic in it has been used to cast a Spell to banish this Curse into the Shadow Realm of MCP and Forge ModLoader.